ref(compiler): consolidate PHP-builtin return-type tables into PhpFunctionReturnTypes
The analyzer's PurePhpFunctionReturnTypes and the emitter's
KnownPhpFunctionReturnTypes had drifted apart. One class now holds a
strict band (exactly one primitive on every non-throwing input; safe to
persist on bindings and fn returns) and an operand-only widening band
(pow: int|float is operator-compatible with every spliced native op and
Phel = is already strict across int/float).

json_encode, hex2bin and str_replace are dropped entirely: their
string|false / string|array returns could splice false into a native
concat, emitting "" where the dynamic str path renders "false".

Related to #2768
