+++ CHANGELOG.md
@@ -40,6 +40,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- String concatenation no longer specializes over `php/json_encode`, `php/hex2bin`, or `php/str_replace`: their `string|false` / `string|array` returns could splice `false` into a native `.` concat and emit `""` where the dynamic `str` path renders `"false"`. The two compiler return-type tables are consolidated into one `PhpFunctionReturnTypes` with an explicit strict band (persistable tags) and a documented operand-only widening (`pow`) (#2768)
 - `phel test --coverage`: xdebug is only used as the coverage driver when `coverage` is among its active modes; otherwise the command fails fast with a hint to re-run with `XDEBUG_MODE=coverage` instead of emitting PHP warnings and an empty report (#2764)
 - Self-recursion inside a def whose init wraps the fn (e.g. `(def f (wrap (fn [n] ... (f ...))))`) no longer emits the `$this` self-call shortcut — the fn compiles to a plain closure there, so the shortcut invoked an unrelated object ("Object of type ... is not callable"); recursion now routes through the registry. Direct `(def f (fn ...))` inits keep the fast path
