+++ tests/php/Integration/Fixtures/Call/str-concat-conditional-return-stays-dynamic.test
@@ -1,0 +1,14 @@
+--PHEL--
+(fn [x] (str "a" (php/json_encode x)))
+(fn [x] (str "b" (php/hex2bin x)))
+(fn [x] (str "c" (php/bin2hex x)))
+--PHP--
+(function($x): string {
+  static $__phel_call_0;
+  return ($__phel_call_0 ??= \Phel::getDefinition("phel.core", "str"))->__invoke("a", json_encode($x));
+});(function($x): string {
+  static $__phel_call_0;
+  return ($__phel_call_0 ??= \Phel::getDefinition("phel.core", "str"))->__invoke("b", hex2bin($x));
+});(function($x): string {
+  return ("c" . bin2hex($x));
+});
